// recaudacion/2.Codigo/frontend/src/app/core/models/pedidopecosa.ts
export class PedidoPecosa {
    anioEje: number;
    ejecutora: string;
    tipoBien: string;
    numeroPecosa: number;
    fechaPecosa: any;
    nombreAlmacen: string
    motivoPedido: string;
    nombreMarca: string;
    codigoItem: string;
    nombreItem: string;
    nombreUnidad: string;
    clasificador: string;
    nombreClaficador: string;
    cantidadAtendida: number;
    cantidadAprobada: number;
    precioUnitario: number;
    valorTotal: number;
    index:number;
}